requirements.txt
Flask>=3.0.0
flask-socketio>=5.3.0
qrcode[pil]>=7.4.2
python-socketio>=5.10.0
Pillow>=11.0.0
